ENH: New approach for the simplified player API.
The general idea is that we give up on the object oriented interface and
instead have the user supply a move function that receives a
universe/state and a storage dict.

The universe/state object should contain all necessary information to
evaluate a state. The representation will be kept as simple as possible
(a simple print should show all available information to help with
observability) and we will export all the previous helper methods on
AbstractPlayer as functions.

Additionally, we initialise an empty storage dict which we pass to the
move function. The storage dict can be used to share data between the
individual players and rounds and is persistent for a whole game.