ENH: First draft for error-handling capabilities in the zmq protocol
pelita_player will now intercept when load_team fails and return an
error message object to the zmq socket. The ZMQConnection has been
adapted to close itself and raise a connection error on retrieval.

As a quick fix, the error message is returned from set_initial as the
team name so that it can be shown in the UI.

Further updates should make this a bit more flexible and refined.